Switch to a window-level capture handler
Some bookmarklets are hosted on pages that bind a capturing event
listener to a higher-level element to prevent the accidental use of
the bookmarklet on its hosting page.

Move all of our click handling into a single capturing handler on
the window object (the highest level).

This should allow us to usurp any capturing handlers installed at
a *lower* level. There is nothing we can do if there is a prior
window-level capturing handler that stops propagation (it will run
first and prevent our handler from running).